+++ cli-tool/lib/overrides.rb
@@ -1,3 +1,5 @@
-def released?
-  false
+module Taylor
+  def self.released?
+    false
+  end
 end