+++ src/workarounds/mingw.cpp
@@ -25,7 +25,7 @@ workarounds_mingw_msg_dontwait(mrb_state* mrb)
   // We don't have this populated on windows, so let's just whack it in at 0.
   // This gets the webserver properly responding to requests on my machine.
   mrb_load_string(mrb, R"(
-    Socket::MSG_DONTWAIT = 0 if windows?
+    Socket::MSG_DONTWAIT = 0 if Taylor::Platform.windows?
   )");
 }
 #endif